+++ data/duke.txt
@@ -1,4 +1,4 @@
 T | 1 | read book
-D | 0 | return book | June 6th
-E | 0 | project meeting | Aug 6th 2-4pm
+D | 0 | return book | 2019-06-06
+E | 0 | project meeting | 2019-08-06
 T | 1 | join sports club
