+++ src/main/java/Deadline.java
@@ -1,19 +1,24 @@
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
+
 public class Deadline extends Task {
 
-    protected String by;
+    protected LocalDate by;
 
-    public Deadline(String description, String by) {
+    public Deadline(String description, LocalDate by) {
         super(description);
         this.by = by;
     }
 
     @Override
     public String toString() {
+        String formattedDate = by.format(DateTimeFormatter.ofPattern("MMM d yyyy"));
         return "[D]" + super.toString() + " (by: "
-                + by + ")";
+                + formattedDate + ")";
     }
 
-    public String getBy() {
+    public LocalDate getBy() {
         return this.by;
     }
 }
