+++ src/main/java/Event.java
@@ -1,19 +1,24 @@
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
+
 public class Event extends Task {
 
-    protected String at;
+    protected LocalDate at;
 
-    public Event(String description, String at) {
+    public Event(String description, LocalDate at) {
         super(description);
         this.at = at;
     }
 
     @Override
     public String toString() {
+        String formattedDate = at.format(DateTimeFormatter.ofPattern("MMM d yyyy"));
         return "[E]" + super.toString() + " (at: "
-                + at + ")";
+                + formattedDate + ")";
     }
 
-    public String getAt() {
+    public LocalDate getAt() {
         return this.at;
     }
 }
